+++ .github/workflows/release.yml
@@ -11,6 +11,6 @@ jobs:
     uses: ASFHyP3/actions/.github/workflows/reusable-release.yml@v0.20.0
     permissions: {}
     with:
-      release_prefix: HyP3 gather-landsat
+      release_prefix: HyP3 akfire-safe
     secrets:
       USER_TOKEN: ${{ secrets.TOOLS_BOT_PAK }}
