+++ .github/workflows/test.yml
@@ -17,6 +17,6 @@ jobs:
     permissions:
       contents: read
     with:
-      local_package_name: hyp3_gather_landsat
+      local_package_name: hyp3_akfire_safe
       python_versions: >-
         ["3.10", "3.11", "3.12"]
