+++ CHANGELOG.md
@@ -11,6 +11,9 @@ and uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Added
 - Added a new workflow `feds` to process fire perimeters using local files.
 
+### Changed
+- Changed name of the plugin to `hyp3-akfire-safe`.
+
 ## [2.0.0]
 
 ### Added
