+++ Dockerfile
@@ -2,13 +2,13 @@ FROM condaforge/mambaforge:latest
 
 # For opencontainers label definitions, see:
 #    https://github.com/opencontainers/image-spec/blob/master/annotations.md
-LABEL org.opencontainers.image.title="HyP3 gather-landsat "
-LABEL org.opencontainers.image.description="Plugin to get landsat scenes from an input location"
+LABEL org.opencontainers.image.title="HyP3 akfire-safe "
+LABEL org.opencontainers.image.description="Plugin with workflows for akfire-safe"
 LABEL org.opencontainers.image.vendor="Alaska Satellite Facility"
 LABEL org.opencontainers.image.authors="username_for_github_actions <email_for@github_actions.com>"
 LABEL org.opencontainers.image.licenses="BSD-3-Clause"
-LABEL org.opencontainers.image.url="https://github.com/ASFHyP3/hyp3-gather-landsat"
-LABEL org.opencontainers.image.source="https://github.com/ASFHyP3/hyp3-gather-landsat"
+LABEL org.opencontainers.image.url="https://github.com/ASFHyP3/hyp3-akfire-safe"
+LABEL org.opencontainers.image.source="https://github.com/ASFHyP3/hyp3-akfire-safe"
 LABEL org.opencontainers.image.documentation="https://hyp3-docs.asf.alaska.edu"
 
 # Dynamic lables to define at build time via `docker build --label`
@@ -36,13 +36,13 @@ USER ${CONDA_UID}
 SHELL ["/bin/bash", "-l", "-c"]
 WORKDIR /home/conda/
 
-COPY --chown=${CONDA_UID}:${CONDA_GID} . /hyp3-gather-landsat/
+COPY --chown=${CONDA_UID}:${CONDA_GID} . /hyp3-akfire-safe/
 
-RUN mamba env create -f /hyp3-gather-landsat/environment.yml && \
+RUN mamba env create -f /hyp3-akfire-safe/environment.yml && \
     conda clean -afy && \
-    conda activate hyp3-gather-landsat && \
-    sed -i 's/conda activate base/conda activate hyp3-gather-landsat/g' /home/conda/.profile && \
-    python -m pip install --no-cache-dir /hyp3-gather-landsat
+    conda activate hyp3-akfire-safe && \
+    sed -i 's/conda activate base/conda activate hyp3-akfire-safe/g' /home/conda/.profile && \
+    python -m pip install --no-cache-dir /hyp3-akfire-safe
 
-ENTRYPOINT ["/hyp3-gather-landsat/src/hyp3_gather_landsat/etc/entrypoint.sh"]
+ENTRYPOINT ["/hyp3-akfire-safe/src/hyp3_akfire_safe/etc/entrypoint.sh"]
 CMD ["-h"]
