+++ environment.yml
@@ -1,4 +1,4 @@
-name: hyp3-gather-landsat
+name: hyp3-akfire-safe
 channels:
   - conda-forge
   - nodefaults
