+++ requirements-static.txt
@@ -1,3 +1,3 @@
 git+https://github.com/Earth-Information-System/fireatlas.git
-ruff==0.14.1
+ruff==0.14.3
 mypy==1.18.2
