+++ src/hyp3_akfire_safe/aux/VIIRS_Global_flaring_d.7_slope_0.029353_2017_web_v1.csv
