--- src/hyp3_gather_landsat/etc/entrypoint.sh
@@ -1,4 +1,0 @@
-#!/bin/bash --login
-set -e
-conda activate hyp3-gather-landsat
-exec python -um hyp3_gather_landsat "$@"
