+++ tests/test_entrypoints.py
@@ -1,3 +1,3 @@
-def test_hyp3_gather_landsat(script_runner):
-    ret = script_runner.run(['python', '-m', 'hyp3_gather_landsat', '-h'])
+def test_hyp3_akfire_safe(script_runner):
+    ret = script_runner.run(['python', '-m', 'hyp3_akfire_safe', '-h'])
     assert ret.success
