+++ tests/test_feds.py
@@ -1,7 +1,7 @@
 import shutil
 from pathlib import Path
 
-from hyp3_gather_landsat import feds
+from hyp3_akfire_safe import feds
 
 
 def test_get_product_name():
